// apps/web/src/app/api-client/models/conformite-entity.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PolitiqueRefEntity } from '../models/politique-ref-entity';
import { SolutionRefEntity } from '../models/solution-ref-entity';
export interface ConformiteEntity {

  /**
   * Commentaire associé au statut de conformité.
   */
  commentaire?: string | null;

  /**
   * Date de création.
   */
  createdAt: string;

  /**
   * Identifiant de la conformité.
   */
  id: string;

  /**
   * Politique de gouvernance évaluée.
   */
  politique: PolitiqueRefEntity;

  /**
   * Identifiant de la politique de gouvernance évaluée.
   */
  politiqueId: string;

  /**
   * Solution évaluée.
   */
  solution: SolutionRefEntity;

  /**
   * Identifiant de la solution évaluée.
   */
  solutionId: string;

  /**
   * Statut de conformité à la politique.
   */
  statut: 'CONFORME' | 'NON_CONFORME' | 'A_EVALUER';

  /**
   * Date de dernière mise à jour.
   */
  updatedAt: string;
}